Cleaned up redundant logging.
