+++ org/diceresearch/nebula/veracity_detection/predictions.py
@@ -1,4 +1,3 @@
-import logging
 import threading
 
 import numpy as np
@@ -7,6 +6,8 @@ import torch
 
 import settings, orchestrator, databasemanager
 from veracity_detection.aggregation import AggregationProcessor
+
+from org.diceresearch.nebula.exception_handling.exception_utils import log_exception
 
 
 def aggregate(json, type):
@@ -50,6 +51,4 @@ def predict(json, identifier):
         thread = threading.Thread(target=orchestrator.goNextLevel, args=(identifier,))
         thread.start()
     except Exception as e:
-        logging.exception(e)
-        databasemanager.update_step(settings.results_table_name, settings.status, settings.error, identifier)
-        databasemanager.update_step(settings.results_table_name, settings.error_msg, str(e), identifier)
+        log_exception(e, identifier)
